rename, extension and comment changes.
